+++ submissions/Ant-C-tech/dom-api-task/index.html
@@ -113,7 +113,7 @@
         </p>
     </footer>
 
-    <script defer src="./js/app.js"></script>
+    <script defer type="module" src="./js/app.js"></script>
 </body>
 
 </html>
