+++ submissions/Ant-C-tech/dom-api-task/js/content.js
@@ -1,0 +1,703 @@
+'use strict'
+
+const content = {
+    main: {
+        title: 'What do you know about Ukraine?',
+        subtitle: 'Test yourself!',
+        text: [
+            `Ukraine, country located in eastern Europe, the second largest on the continent after Russia. The
+                    capital is Kyiv
+                    (Kiev), located on the Dnieper River in north-central Ukraine.
+                Ukraine is bordered by Belarus to the north, Russia to the east, the Sea of Azov and the Black Sea to
+                    the south,
+                    Moldova
+                    and Romania to the southwest, and Hungary, Slovakia, and Poland to the west. In the far southeast,
+                    Ukraine is
+                    separated
+                    from Russia by the Kerch Strait, which connects the Sea of Azov to the Black Sea.`,
+            `Ukraine lies in a temperate climatic zone influenced by moderately warm, humid air from the Atlantic
+                    Ocean. Winters
+                    in
+                    the west are considerably milder than those in the east. In summer, on the other hand, the east
+                    often experiences
+                    higher
+                    temperatures than the west.`,
+            `By the turn of the 21st century, ethnic Ukrainians made up more than three-fourths of the population.
+                    Russians continue
+                    to be the largest minority, though they now constitute less than one-fifth of the population. The
+                    remainder of the
+                    population includes Belarusians, Moldovans, Bulgarians, Poles, Hungarians, Romanians, Roma
+                    (Gypsies), and other groups.`
+        ],
+        media: [
+            {
+                src: './img/main-content/Flag_of_Ukraine.jpg',
+                descript: 'Flag of Ukraine',
+            },
+            {
+                src: './img/main-content/1598538980.jpg',
+                descript: 'Coat of arms of Ukraine',
+            },
+            {
+                src: './img/main-content/Kyiv_logo_english.jpg',
+                descript: 'Touristic logo of Kyiv',
+            },
+        ],
+    },
+
+    'UA-09': {
+        title: 'Luhansk Region',
+        subtitle: 'Population is estimated as 2,135,913 (2020)',
+        text: [
+            `Luhansk Region is located in eastern Ukraine. The area of the region (26,700 km²), comprises about 4.42% of the total area of the country. Its longitude from north to south is 250 km, from east to west – 190 km. The region is located in the valley of Siversky Donets which splits the region approximately in half. The southern
+                portion of the region is elevated by the Donetsk Ridge which is located closer to the southern border. The highest point is Mohyla Mechetna (367 m (1,204 ft)) which is the highest point of Donetsk Ridge.`,
+            `The population is largely Russian-speaking, although ethnic Ukrainians constitute a majority (58.0%). Among the
+                minorities are native Russians (39.1%), Belarusians (0.8%), and others (1.4%).`,
+            `Luhansk is the administrative center of the Luhansk Region. Its population estimated as 401,297. Also the province is primarily divided into 18 raions (districts), and 37 cities.`
+        ],
+        media: [
+            {
+                src: './img/ua-09/757298.jpg',
+                descript: 'Landscape of Luhansk Region',
+            },
+            {
+                src: './img/ua-09/Coat_of_Arms_Luhansk_Oblast.jpg',
+                descript: 'Coat of Arms of Luhansk Region',
+            },
+            {
+                src: './img/ua-09/b65b92c-299.jpg',
+                descript: 'Luhansk',
+            },
+        ],
+    },
+
+    'UA-63': {
+        title: 'Kharkiv Region',
+        subtitle: 'Population is estimated as 2,658,461 (2020)',
+        text: [
+            `Kharkiv Region is the main part of historical-cultural and geographical region of Sloboda Ukraine.
+                    The massive settling of this territory with Ukrainian Cossacks, villages and migrants from Russia
+                    started in ХVІІ century. The area of Kharkiv Region is 31,4 thousand sq.m. The number of population
+                    is 2,95 million people. There are 27
+                    administrative districts and 20 cities in the region.`,
+            `Kharkiv Region is the leading industrial, agricultural, scientific and cultural center of Ukraine.
+                    The share of the
+                    regional industrial production makes 5,9% of the Ukrainian industrial production volume.`,
+            `Kharkiv Region is the Slobozhanshchyna culture center with rich traditions having deep historical roots. Such world-wide
+                    known figures as philosopher Grygory Skovoroda, writer Grygory Kvitka-Osnovyanenko, painter Illya Ryepin, producer Les
+                    Kurbas and architect Beketov lived and worked here. Today the culture admirers are welcome to theatres, regional
+                    philharmonic society, House of organ and chamber music, circus, museums and about 1000 libraries.`
+        ],
+        media: [
+            {
+                src: './img/ua-63/400x280-HBVfD.jpg',
+                descript: 'Kharkiv',
+            },
+            {
+                src: './img/ua-63/COA_of_Kharkiv_Oblast.png',
+                descript: 'Coat of Arms of Kharkiv Region',
+            },
+            {
+                src: './img/ua-63/400x280--s_4l.jpg',
+                descript: 'Krasnokutsk',
+            },
+        ],
+    },
+
+    'UA-14': {
+        title: 'Donets\'k Region',
+        subtitle: 'Population is estimated as 4,131,808 (2020)',
+        text: [
+            `Powerful industrial Donets\'k Region is situated in the steppe zone of the south-eastern part of Ukraine. It was established on July 17, 1932. The area is 26.5 thousand square km, or 4.4% of the total territory of Ukraine. There are 28 cities and 17 districts of regional value, 134 settlements of urban type and 1122 rural settlements in the regional administrative-territorial structure.`,
+            `The regional economic potential is 822 industrial, 16 thousand small and almost 600 agricultural enterprises of all forms of property, over 500 specialized construction organizations and 2132 farm facilities.`,
+            `The regional economy is served by 137 banks and banking institutions, 17 insurance companies and 24 stock exchanges. The interests of enterprises and organizations of production, commercial, agroindustrial, scientific-technical, insurance and bank-financial spheres of the region in Ukraine and abroad are represented by the Donetsk Chamber of Commerce and Industry.`
+        ],
+        media: [
+            {
+                src: './img/ua-14/caption.jpg',
+                descript: 'Sviatohirsk',
+            },
+            {
+                src: './img/ua-14/200px-Insigne_Donetzensis.jpg',
+                descript: 'Coat of Arms of Donetsk Region',
+            },
+            {
+                src: './img/ua-14/images.jpg',
+                descript: 'Donets\'k',
+            },
+        ],
+    },
+
+    'UA-23': {
+        title: 'Zaporizhzhya Region',
+        subtitle: 'Population is estimated as 1,687,401 (2020)',
+        text: [
+            `Zaporizhzhya Region is one of the impotant components of economic potential of Ukraine. The oblast covers an area of 27,183 km2 (10,495 sq mi). The area corresponding approximately to the modern Zaporizhzhia Oblast - according to Herodotus - was called in the antiquity as the land of Gerrhos. More than 70.8% people considered themselves Ukrainians, while 24.7% were Russians, the rest were of assorted nationalities including Bulgarians (1.4%), Belarusians (0.7%), and others (1.6%).`,
+            `The ninth place in Ukraine by number of population, first place by industrial production volume per capita and third place by general industrial production volume in operating prices (after Donetsk and Dnipropetrovs'k regions) is the economic rating of Zaporizhzhya Region today`,
+            `Zaporizhzhya Region thanks to the powerful industrial and scientific and technical potential, presence of natural resources, home energy resources, availability of markets of Ukraine and the CIS, well-developed transport infrastructure and banking system is one of the most investment attractive regions.`
+        ],
+        media: [
+            {
+                src: './img/ua-23/123.jpg',
+                descript: 'The National Reserve «Khortytsia»',
+            },
+            {
+                src: './img/ua-23/Coat_of_arms_of_Zaporizhia_Oblast.jpg',
+                descript: 'Coat of Arms of Zaporizhzhya Region',
+            },
+            {
+                src: './img/ua-23/images.jpg',
+                descript: 'Zaporizhzhya',
+            },
+        ],
+    },
+
+    'UA-12': {
+        title: 'Dnipropetrovs\'k Region',
+        subtitle: 'Population is estimated as 3,176,648 (2020)',
+        text: [
+            `Dnipropetrovsk Region is one of the most perspective, industrial and cultural well-developed centers of Ukraine, potential and reliable partner of the European Community and its business structures. Today Dnipropetrovs'k Region comprises 22 administrative districts, 21 cities, 47 settlements of urban type and 1440 rural settlements.`,
+            `Dnipropetrovs'k Region has powerful production potential. The share of Dnipropetrovs'k Region in the national industrial production makes 15,7% of general volume, besides the most minerals of Ukraine are extracted here. The regional enterprises provide 82% of iron-stone of the national production, 100% of manganese ore, 36% of cast-iron, 33.6% of steel, 73.5% of steel pipes, 13% of finished coal, 15% of mineral fertilizers and 24% of tires. `,
+            `Sending its development on the future, Dnipropetrovs'k Region develops science and culture. Today about 160 thousand students are studying in higher educational establishments of the region. The European choice of Ukraine is that pointer for Ukrainian peoples, which means construction of civil society step by step and socially oriented market economy.`
+        ],
+        media: [
+            {
+                src: './img/ua-12/img-20170918-172009-497.jpg',
+                descript: 'Dnipro',
+            },
+            {
+                src: './img/ua-12/Large_Coat_of_Arms_of_Dnipropetrovsk_Oblast.jpg',
+                descript: 'Coat of Arms of Dnipropetrovsk Region',
+            },
+            {
+                src: './img/ua-12/123.jpg',
+                descript: 'Dnipro',
+            },
+        ],
+    },
+
+    'UA-53': {
+        title: 'Poltava Region',
+        subtitle: 'Population is estimated as 1,386,978 (2020)',
+        text: [
+            `The Poltava land is situated in the Middle Dnieper riverside and is one of those historical regions having become the core of the Ukrainian statehood. Its historic and cultural properties, natural riches and powerful national economic complex play an important role in the state building of independent Ukraine.`,
+            `The mathematician Mykhaylo Ostrogradskiy, surgeon Mykola Sklifosovskiy, scientist-ground expert Vasyl Dokuchayev and many other prominent scientists with the world name went to the glory from the Poltava land. The inhabitants of Poltava Region are proud that the life and creation of such global mental giants as Volodymyr Vernadskiy and Volodymyr Korolenko and world-known teachers Vasyl Sukhomlynskiy and Anton Makarenko is connected with the Poltava land.`
+        ],
+        media: [
+            {
+                src: './img/ua-53/aerial_view_on_kruhla_square_poltava_ukraine.jpg',
+                descript: 'Poltava',
+            },
+            {
+                src: './img/ua-53/Coat_of_Arms_of_Poltava_Oblast.jpg',
+                descript: 'Coat of Arms of Poltava Region',
+            },
+            {
+                src: './img/ua-53/poltava-central-square.jpg',
+                descript: 'Poltava',
+            },
+        ],
+    },
+
+    'UA-59': {
+        title: 'Sumy Region',
+        subtitle: 'Population is estimated as 1,068,247 (2020)',
+        text: [
+            `Sumy Region has the highly developed diversified industrial complex with 266 large and medium enterprises and more than 800 small enterprises. The machine-building and metal-working are leading branches of the regional industry in the conditions of market transformation.`,
+            `The extractive industry of Sumy Region is well developed. There is extracted 40% of the national volume of oil in the region. The extraction of high-quality quartzites (content SiO2 – 98-99%) used for production of silicon and ferro-alloys is at high level.`,
+            `Sumy Region has been significantly changed and its cities and villages became more beautiful and merrier. It is pleasantly to mark that there is a particle of labour of each resident of the region in these changes.`
+        ],
+        media: [
+            {
+                src: './img/ua-59/108987941.jpg',
+                descript: 'Sumy',
+            },
+            {
+                src: './img/ua-59/214px-Coat_of_Arms_of_Sumy_Oblast.jpg',
+                descript: 'Coat of Arms of Sumy Region',
+            },
+            {
+                src: './img/ua-59/alexander-gymnasium-sumy.jpg',
+                descript: 'Sumy',
+            },
+        ],
+    },
+
+    'UA-43': {
+        title: 'Autonomous Republic of Crimea',
+        subtitle: 'Population is estimated as 2,033,700 (2001)',
+        text: [
+            `The territory of Crimean Autonomous Republic is 26,1 thousand square km. The Crimea borders with Kherson and Zaporizhzhya regions of Ukraine and with Krasnodar Region of the Russian Federation. The State frontier length is 821 km and it can be crossed in nine check points. Crimean Autonomous Republic is divided into 14 districts by administrative structure.`,
+            `The land area of the Autonomous Republic of Crimea and Sevastopol, their internal waters are defined under the Law of Ukraine "On the rights and freedoms of citizens and legal regime in the temporarily occupied territory of Ukraine" as temporarily occupied territories as a result of military aggression of the Russian Federation.`
+        ],
+        media: [
+            {
+                src: './img/ua-43/250px-Flag_of_the_Crimean_Tatar_people.jpg',
+                descript: 'Crimean Tatar\'s Flag',
+            },
+            {
+                src: './img/ua-43/Emblem_of_Crimea.jpg',
+                descript: 'Coat of Arms of Crimea',
+            },
+            {
+                src: './img/ua-43/150320151120_russian_army_crimea_624x351_reuters.jpg',
+                descript: 'Russian Army',
+            },
+        ],
+    },
+
+    'UA-40': {
+        title: 'Autonomous Republic of Crimea',
+        subtitle: 'Population is estimated as 2,033,700 (2001)',
+        text: [
+            `The territory of Crimean Autonomous Republic is 26,1 thousand square km. The Crimea borders with Kherson and Zaporizhzhya regions of Ukraine and with Krasnodar Region of the Russian Federation. The State frontier length is 821 km and it can be crossed in nine check points. Crimean Autonomous Republic is divided into 14 districts by administrative structure.`,
+            `The land area of the Autonomous Republic of Crimea and Sevastopol, their internal waters are defined under the Law of Ukraine "On the rights and freedoms of citizens and legal regime in the temporarily occupied territory of Ukraine" as temporarily occupied territories as a result of military aggression of the Russian Federation.`
+        ],
+        media: [
+            {
+                src: './img/ua-43/250px-Flag_of_the_Crimean_Tatar_people.jpg',
+                descript: 'Crimean Tatar\'s Flag',
+            },
+            {
+                src: './img/ua-43/Emblem_of_Crimea.jpg',
+                descript: 'Coat of Arms of Crimea',
+            },
+            {
+                src: './img/ua-43/150320151120_russian_army_crimea_624x351_reuters.jpg',
+                descript: 'Russian Army in Crimea',
+            },
+        ],
+    },
+
+    'UA-65': {
+        title: 'Kherson Region',
+        subtitle: 'Population is estimated as 1,027,913 (2020)',
+        text: [
+            `There are 18 administrative districts, 3 cities of regional submission and 658 settlements in the region.
+            Kherson Region is washed by the Black Sea and Sea of Azov. On the east it borders with Zaporizhzhya Region as well as north-west – Mykolayiv Region, north – Dnipropetrovs'k Region and south – Crimean Autonomous Republic. The state frontier with the length of 458 km passes on the region's territory including 350 km of the Black Sea and 108 km of the Sea of Azov. The area of the region is 28,461 km2.`,
+            `About 61.5% or 745,400 people live in urban areas of the Oblast and 38.5% or 467,600 people live in agricultural centers/villages. This is considered to be the 'fruit basket' of the country, as many of the oblast's agricultural production is dispersed throughout the country, with production peaking during the summer months.`
+        ],
+        media: [
+            {
+                src: './img/ua-65/ac63c874506f44c8fb3e89081a189f5a.jpg',
+                descript: 'Dzharylhach Island',
+            },
+            {
+                src: './img/ua-65/Coat_of_Arms_of_Kherson_Oblast.jpg',
+                descript: 'Coat of Arms of Kherson Region',
+            },
+            {
+                src: './img/ua-65/Travel-Ukraine-Today-Kherson-13.jpg',
+                descript: 'Kherson Region',
+            },
+        ],
+    },
+
+    'UA-74': {
+        title: 'Chernihiv Region',
+        subtitle: 'Population is estimated as 991,294 (2020)',
+        text: [
+            `The capital city of Chernihiv has known human settlement for over 2,000 years, according to archaeological excavations. The Chernihiv Oblast comprises a very important historical region, notable as early as the Kievan Rus' period, when the cities of Chernihiv and Novhorod-Siverskyi were frequently mentioned. The city of Chernihiv was the second most important Ukrainian city during the Rus' period of Ukrainian history, often serving as a major regional capital. Danylo of Chernihiv wrote of his pilgrimage to Jerusalem during this era. The numerous architectural monuments of the city bear witness to the invasions suffered, including those by the Tatars/Mongols, Lithuanians, Poles, and Nazis. On the territory of the region took place numerous tragic events such as Battle of Kruty, Tragedy of Kryukivka, and many others.`,
+            `Occupying 5% of the territory, where 2.6% of population of Ukraine lives, the region provides production 70% of total Ukrainian output of woolens, 80% of wallpapers, 34% of chemical fibres and 13% of extraction of oil. Chernihiv Region, unique in Ukraine, produces cord fabric, fire trucks, protein coats, pianos, separate types of equipment for branches of agroindustrial complex.`
+        ],
+        media: [
+            {
+                src: './img/ua-74/chernihiv-ukraine-33.jpg',
+                descript: 'Chernihiv',
+            },
+            {
+                src: './img/ua-74/263px-Coat_of_Arms_of_Chernihiv_Oblast.jpg',
+                descript: 'Coat of Arms of Chernihiv Region',
+            },
+            {
+                src: './img/ua-74/images.jpg',
+                descript: 'Chernihiv',
+            },
+        ],
+    },
+
+    'UA-32': {
+        title: 'Kyiv Region',
+        subtitle: 'Population is estimated as 1,781,044 (2020)',
+        text: [
+            `Kyiv land is streched in the centre of Ukraine, on both picturesque banks of the mighty Dnieper River. The archaeological sights of Trypillya, Chernyakhiv, Lebedivska and Sofiyivska cultures testify to ancient history of this land and the settlements of the region gave the proper names to them.`,
+            `Its territory is 28.4 thousand square km (4.7% of the total Ukrainian area). The number of population is 1810,5 thousand people (3.7% of the total number of Ukraine). The urban population makes 58.9% as well as rural – 41.1% of the total number. Kyiv Region borders with Zhytomyr, Vinnytsya, Cherkasy, Poltava and Chernihiv regions, and also with Gomel Region of the Republic Byelorussia. Kyiv Region comprises 25 districts and 24 cities, including 11 cities of regional submission, 30 settlements and 1220 villages.`,
+            `Kyiv land since ancient times is famous by works of decorative-trade art, created by talented masters. There are famous names of Catherine Bilokur in the world, Maria Pryymachenko, Hanna Veres and Ivan Prykhodka.`
+        ],
+        media: [
+            {
+                src: './img/ua-32/3c7adcc87e0eb8f8795f2f5b4eb0e19d.jpg',
+                descript: 'Kyiv Region',
+            },
+            {
+                src: './img/ua-32/366px-Herb_Kyivskoi_oblasti_1.jpg',
+                descript: 'Coat of Arms of Kyiv Region',
+            },
+            {
+                src: './img/ua-32/landscape-park-buky-kyiv-region-ukraine-1.jpg',
+                descript: 'Kyiv Region',
+            },
+        ],
+    },
+
+    'UA-30': {
+        title: 'The City of Kyiv',
+        subtitle: 'Population is estimated as 2,847,200 (2013)',
+        text: [
+            `Kyiv is the political center of the country, where the residence of country leader – the President of Ukraine is located as well as Supreme Soviet of Ukraine, Cabinet of Ministers of Ukraine, Constitutional Court of Ukraine, representatives of international organizations in Ukraine. It is the spiritual, cultural, historical and scientific-educational center of Ukraine.`,
+            `The capital confidently leads among other regions of the country by practically all basic industries. The share of Kyiv in the gross value added of Ukraine annually grows and exceeds 11% today. Thus the number of city population makes only 5.4% of Ukrainian people.`,
+            `Geographically, Kyiv is located on the border of the Polesia woodland ecological zone, a part of the European mixed woods area, and the East European forest steppe biome. However, the city's unique landscape distinguishes it from the surrounding region.`
+        ],
+        media: [
+            {
+                src: './img/ua-30/17-07-02-Maidan_Nezalezhnosti_RR74377-PANORAMA.jpg',
+                descript: 'Maidan Nezalezhnosti, Kyiv',
+            },
+            {
+                src: './img/ua-30/449px-COA_of_Kyiv_Kurovskyi.jpg',
+                descript: 'Coat of Arms of Kyiv',
+            },
+            {
+                src: './img/ua-30/123.jpg',
+                descript: 'Bankova str, Kyiv',
+            },
+        ],
+    },
+
+    'UA-71': {
+        title: 'Cherkasy Region',
+        subtitle: 'Population is estimated as 1,192,137 (2020)',
+        text: [
+            `With 20,900 km², the Cherkasy oblast is the 18th largest oblast of Ukraine, comprising about 3.5% of the area of the country. The south flowing Dnieper River with the hilly western bank and the plain eastern bank divides the oblast into two unequal parts. The larger western part belongs to the Dnieper Upland. The low-lying eastern part of the oblast used to be subject to the frequent Dnieper flooding before the flow of the river became controlled by multiple dams of Hydroelectric Power Plants constructed along the river in the 20th century.`,
+            `Cherkasy Region is rich in minerals and the most found and developed are non-metallic resources. Over one hundred deposits of the brick-tiled raw material are known that enables to produce roofing and facing materials.`,
+            `Cherkasy land gave life and accepted in its cuddles prophet of Ukraine Taras Shevchenko. The state mind and military talent of hetman Bogdan Khmelnytsky has been raised in Cherkasy land and the local people has firstly glorified his exploits, and exploits of his associates as Ivan Bogun and Maxim Kryvonos. `
+        ],
+        media: [
+            {
+                src: './img/ua-71/123.jpg',
+                descript: 'Cherkasy Region',
+            },
+            {
+                src: './img/ua-71/Coat_of_Arms_of_Cherkasy_Oblast.jpg',
+                descript: 'Coat of Arms of Cherkasy Region',
+            },
+            {
+                src: './img/ua-71/cherkassy-ukraine-oblast-views-1.jpg',
+                descript: 'Cherkasy Region',
+            },
+        ],
+    },
+
+    'UA-35': {
+        title: 'Kirovograd Region',
+        subtitle: 'Population is estimated as 965,756 (2017)',
+        text: [
+            `Kirovograd Region is situated in the centre of Ukraine, in the interfluve between the Dnieper and South Buh rivers. It confidently became on the way of dynamic economic and social development. The regional industrial complex, where 292 enterprises operate, comprises 10 leading industries. The economy of Kirovohrad Region gets more social orientation and the production of commodities of public consumption grows by passing ahead rates. In the structure of commodities of public consumption, produced by the regional enterprises, the food stuffs make 77.5% as well as non food – 16.9%; wine and alcoholic products – 5.6%.`,
+            `Kirovohrad Region, taking into account motion of Ukraine into the European Community, has all chances to become one of the main suppliers of high-quality and mainly ecologically friendly products of feed on the foreign market. The skilled specialists and industrial potential, favourable geographical location are to bring our Ukrainian steppe pearl to the number of the most perspective regions of Ukraine.`
+        ],
+        media: [
+            {
+                src: './img/ua-35/6831.jpg',
+                descript: 'Kropyvnytskyi',
+            },
+            {
+                src: './img/ua-35/416px-Coat_of_Arms_of_Kirovohrad_Oblast.jpg',
+                descript: 'Coat of Arms of Kirovograd Region',
+            },
+            {
+                src: './img/ua-35/about-ukraine-770x330.jpg',
+                descript: 'Kropyvnytskyi',
+            },
+        ],
+    },
+
+    'UA-51': {
+        title: 'Odesa Region',
+        subtitle: 'Population is estimated as 2,377,230 (2017)',
+        text: [
+            `The Odesa land is the territory of the boundles Ukrainian steppe and genial Black Sea. It is situated on the crossroads of the important international ways (the features of geographic position of the region conditioned the entrance of its territory two of nine routes of international transport corridors); Odesa Region is a part of Ukrainian sea front as well as seacoast and boundary region. The state Ukrainian frontier with Romania and Moldova passes on the region’s territory.`,
+            `The area of Odesa Region makes 33.3 thousand sq. km (5.5% of the territory of Ukraine) that almost equals the sizes of the territory of Republic of Moldova. The population of Odesa Region is 2448.2 thousand people. The region is multinational: Ukrainians, Russians, Jewries, Moldovans and other people live here.`,
+            `The strategy of development of tourism and recreation economy in the region is directed on creation of highly profitable industry that is to cover demands of internal and international tourism, strengthening of inter-relations between tourism and other spheres of regional national economic complex, saving of monuments of history and culture, level of environment protection activity and attraction of investments in development of tourism infrastructure.`
+        ],
+        media: [
+            {
+                src: './img/ua-51/6838.jpg',
+                descript: 'Odesa Region',
+            },
+            {
+                src: './img/ua-51/Coat_of_Arms_of_Odesa_Oblast.jpg',
+                descript: 'Coat of Arms of Odesa Region',
+            },
+            {
+                src: './img/ua-51/od2.jpg',
+                descript: 'Odesa',
+            },
+        ],
+    },
+
+    'UA-05': {
+        title: 'Vinnytsya Region',
+        subtitle: 'Population is estimated as 1,545,416 (2017)',
+        text: [
+            `The region is situated in the central part of the Right-Bank Ukraine and occupies 4.5% of its territory (26.5 thousand square km). There are 17 cities, 30 settlements of urban type, 27 districts and 1467 villages in the region.`,
+            `Vinnytsya Region is situated in the forest-steppe zone within the limits of the Volyn-Podil and Dnieper sublimity. From the north-west to the southeast the wide valley of the South Buh River is laying. Besides, 230 rivers with the length more than 10 kilometres each flow on the territory of Vinnytsya Region. They belong to the basins of the South Buh, Dnestr and Dnieper.`,
+            `With Vinnytsya land there are linked the names of Ustym Karmelyuk, Oleksandr Suvorov, Ivan Kotlyarevsky, Lesya Ukrayinka, Mark Vovchok, Yaroslav Ivashkevych, Petro Tchaikovsky, Mykola Leontovych, Mykhaylo Grushevsky and Vasyl Stus. Vinnytsya land presented many prominent figures to our culture, science and art. `
+        ],
+        media: [
+            {
+                src: './img/ua-05/vinnitsya1-1.jpg',
+                descript: 'Vinnytsya',
+            },
+            {
+                src: './img/ua-05/263px-Coat_of_Arms_of_Vinnytsa_Oblast.jpg',
+                descript: 'Coat of Arms of Odesa Region',
+            },
+            {
+                src: './img/ua-05/unnamed.jpg',
+                descript: 'Vinnytsya Region',
+            },
+        ],
+    },
+
+    'UA-18': {
+        title: 'Zhytomyr Region',
+        subtitle: 'Population is estimated as 1,208,212 (2017)',
+        text: [
+            `Zhytomyr Region (northern part in the area of Polissya and southern part in the forest-steppe) is located in central part of the East-European plain in the north of the Right Bank Ukraine in two natural zones. This is one of the largest regions in Ukraine. Its area makes 29.9 thousand square km.`,
+            `The economy of Zhytomyr Region mostly deals with mining of granite and other construction stone, forestry, agriculture and various machinery manufacturing.The northern part of the province is highly affected by the Chernobyl disaster: some of the towns and raions are devastated and are included in the Chernobyl zone, while others are prohibited from producing their own agriculture.`,
+            `Zhytomyr residents see their happy future in the inspired labour for the good of their land.`
+        ],
+        media: [
+            {
+                src: './img/ua-18/2469.jpg',
+                descript: 'Zhytomyr',
+            },
+            {
+                src: './img/ua-18/396px-Coat_of_Arms_of_Zhytomyr_Oblast.jpg',
+                descript: 'Coat of Arms of Zhytomyr Region',
+            },
+            {
+                src: './img/ua-18/st-clare-church-horodkivka-zhytomyr-region-ukraine-1.jpg',
+                descript: 'Horodkivka',
+            },
+        ],
+    },
+
+    'UA-77': {
+        title: 'Chernivtsi Region',
+        subtitle: 'Population is estimated as 901,632 (2020)',
+        text: [
+            `Chernivtsi Region is situated on the crossroads throughfares between the Central, Southern and Eastern Europe in the foot-hill of the Carpathian. The regional area is 8,1 thousand square km. The region comprises 11 districts, 11 cities, including two cities of regional submission, 8 settlements of urban type and 397 rural settlements. The population representative of almost 80 nationalities and ethnic groups.`,
+            `In the region there are 75 rivers longer than 10 kilometers. The largest rivers are the Dnister (290 km, in the Oblast), Prut (128 km, in the Oblast) and Siret (113 km, in the Oblast).`,
+            `One of priority industries of economy of this land is tourism. The geological and geomorphological formations of Northern Bukovyna have the large cognitive and aesthetical value. The residents of Bukovyna love and can act as host. They are sincerely interested in that the circle of their friends becomes wider.`
+        ],
+        media: [
+            {
+                src: './img/ua-77/800px-Chernivtsi_University.jpg',
+                descript: 'Chernivtsi',
+            },
+            {
+                src: './img/ua-77/Coat_of_Arms_of_Chernivtsi_Oblast.jpg',
+                descript: 'Coat of Arms of Chernivtsi Region',
+            },
+            {
+                src: './img/ua-77/chernovtsy-ukraine-oblast-views-1.jpg',
+                descript: 'Chernivtsi Region',
+            },
+        ],
+    },
+
+    'UA-48': {
+        title: 'Mykolayiv Region',
+        subtitle: 'Population is estimated as 1,119,862 (2020)',
+        text: [
+            `The boundless fields of Pobuzhzhya and Ingul River region, green vineyards and wonderful flowering gardens under bottomless blue sky is Mykolayiv land.
+The agriculture is the leading industry of the regional economy. Mykolayiv Region makes significance payment in strengthening of the country, producing almost seven percents of cereals and sunflower, three percents of gross milk production.`,
+            `Mykolayiv citizens thanks to the labour achievements inscribe worthy pages in the chronicle of native land.`,
+            `There are three professional theatres, 1180 establishments of culture, including 577 clubs in the region. 552 libraries, two state regional museums, three parks of culture and rest, zoo, are functioning in the region.`
+        ],
+        media: [
+            {
+                src: './img/ua-48/nikolaev-ukraine-oblast-views-12.jpg',
+                descript: 'Mykolayiv Region',
+            },
+            {
+                src: './img/ua-48/Coat_of_Arms_of_Mykolaiv_Oblast.jpg',
+                descript: 'Coat of Arms of Mykolayiv Region',
+            },
+            {
+                src: './img/ua-48/nikolaev-ukraine-oblast-views-15.jpg',
+                descript: 'Mykolayiv Region',
+            },
+        ],
+    },
+
+    'UA-68': {
+        title: 'Khmelnytsky Region',
+        subtitle: 'Population is estimated as 1,254,702 (2020)',
+        text: [
+            `Khmelnytsky Region is situated on the Right-Bank Ukraine, in the middle part of Volyn-Podil sublimity. No by chance more than one century age genius Ukrainian poetess Lesya Ukrayinka with love and genuine fascination wrote about that land: «Beautiful Ukraine, Podollya!»`,
+            `The region has inexhaustible natural-raw riches. This is almost 400 deposits of various minerals and the main are follows as saponite clays, phosphorites, apatites, glauconite sands, granite, limestone, sand, gypsum, chalk, flint and, graphite.`,
+            `The cultural life of Khmelnytsky land is rich. Not only in the region but far beyond the land the singing and dancing companies such as Podillya Cossacks, Honta, Horyn and Podolyanchyk are well-known. The musical bands of the Music and Drama Theatre and other artistic establishments renew their repertoire.`
+        ],
+        media: [
+            {
+                src: './img/ua-68/chelnickiu-10.jpg',
+                descript: 'Khmelnytsky',
+            },
+            {
+                src: './img/ua-68/Coat_of_Arms_of_Khmelnytskyi_Oblast.jpg',
+                descript: 'Coat of Arms of Khmelnytsky Region',
+            },
+            {
+                src: './img/ua-68/123.jpg',
+                descript: 'Khmelnytsky',
+            },
+        ],
+    },
+
+    'UA-56': {
+        title: 'Rivne Region',
+        subtitle: 'Population is estimated as 1,152,961 (2020)',
+        text: [
+            `The region is located almost in the middle of the historical region of Volhynia which is indicated on its coat of arms with a white cross on a red background. The relief of the region varies, its northern portion lies at Polesian Lowland, while its southern is located within Volhynian Upland. The highest hills known as Povcha Upland reach over 350 meters. Big portions of the region covered in woodlands.`,
+            `The purposeful activity of the regional executive authorities on creation of favourable business and investment climate in combination with powerful economic, scientific and technical, resource and human potential of the region and its successful geographical location assists to expansion of inter-regional and intergovernmental collaboration volumes, to gradual transition of economic complex and social sphere of the region into the European standards of production, consumption and social security.`
+        ],
+        media: [
+            {
+                src: './img/ua-56/11101130_495252377292747_7920576625895354377_n.jpg',
+                descript: 'Rivne Region',
+            },
+            {
+                src: './img/ua-56/252px-Coat_of_Arms_of_Rivne_Oblast.jpg',
+                descript: 'Coat of Arms of Rivne Region',
+            },
+            {
+                src: './img/ua-56/Tr00100420200228194058963.jpg',
+                descript: 'Rivne Region',
+            },
+        ],
+    },
+
+    'UA-61': {
+        title: 'Ternopil Region',
+        subtitle: 'Population is estimated as 1,038,695 (2020)',
+        text: [
+            `The sources of history of Ternopil Region dates back to the distant times of Kyiv Rus and Galych-Volyn principality. For many centuries the people have been creating the eternal values for the motherland fruitfully preserved traditions, customers and national culture.`,
+            `The characteristic sign of nature of Ternopil land is its variety. The flat central part of the region's territory is surrounded by Kremenets sublimity in the north as well as on the west – Berezhansko-Monastyrsky (Podil) humpbacked mountain, on the east – Avratynska sublimity and on the south it is cut by deep and often canyon-like river valleys of left inflows of the Dniestr River. The central part of the region is crossed from the north-west to the south-east by the Tovtrova chine (Medobory).`,
+            `Its area is 13.8 thousand square km that makes 2.3% of the Ukrainian territories. It stretches from the north to south upon 195 km and from the west to east upon 129 km.`
+        ],
+        media: [
+            {
+                src: './img/ua-61/zalishchyky-resort-ternopil-region-ukraine-1.jpg',
+                descript: 'Zalishchyky',
+            },
+            {
+                src: './img/ua-61/Coat_of_Arms_of_Ternopil_Oblast.jpg',
+                descript: 'Coat of Arms of Ternopil Region',
+            },
+            {
+                src: './img/ua-61/6861.jpg',
+                descript: 'Rivne Region',
+            },
+        ],
+    },
+
+    'UA-26': {
+        title: 'Ivano-Frankivs’k Region',
+        subtitle: 'Population is estimated as 1,368,097 (2020)',
+        text: [
+            `The harmonious fir-trees rise in the sky by green wall. The sweet-scented grass covers a mountain valley by colourful carpet. Wherever you glance over the forest greens everywhere, the Blue Mountains tower over, and between them the rapid rivers and streams discharge it pure waters to the Prut River and further to the wide reaches of the Dnister River.`,
+            `But Ivano-Frankivs’k Region is not only a beauty. Here there are highly developed agriculture and forestry, electric power industry, fuel, chemical and petrochemical, machine-building and metal-working, woodworking, light and food branches of the industry. Here more than 7100 enterprises and 38 500 registered businessmen successfully work.`,
+            `Welcome to Ivano-Frankivs’k Region! Here you will find friends and business partners, will be enriched with new unforgettable impressions, improve your health, take force and enthusiasm to life, labour and creation.`
+        ],
+        media: [
+            {
+                src: './img/ua-26/hc_revato-1987407-86f26cea163ae71180c2000d3a918fae-779992.jpg',
+                descript: 'Ivano-Frankivs’k Region',
+            },
+            {
+                src: './img/ua-26/Ivano-Frankivsk_Oblast-COA.jpg',
+                descript: 'Coat of Arms of Ivano-Frankivs’k Region',
+            },
+            {
+                src: './img/ua-26/Ratush-01.jpg',
+                descript: 'Ivano-Frankivs’k',
+            },
+        ],
+    },
+
+    'UA-21': {
+        title: 'Zakarpattya Region',
+        subtitle: 'Population is estimated as 1,254,614 (2001)',
+        text: [
+            `Zakarpattya is one of the most original corners of Ukraine. This is a region with the unique geopolitical and geographical location (covering the south-west part of the Ukrainian Carpathians and Dunabe lowland). Due to its geographical features Zakarpattya has been for long a convenient junction between the countries of Northern and Southern, Eastern and Western Europe.`,
+            `The territory of Zakarpattya Region is 12.8 thousand square km. By administrative-territorial division this is 13 districts, 10 cities including three cities of regional submission. There are 608 settlements in the region on the whole. `,
+            `But the richest of this land is its hard-working, friendly, wise and talented people. They were famous by trade, generosity and rich of soul for ages. How skilfully they always put in order their dwellings, as want perfection, peace and consent in our general house. Come in this magic corner of emerald forests and mountain streams, in the land of hospitable and benevolent people and you make sure in this.`
+        ],
+        media: [
+            {
+                src: './img/ua-21/avatar_1548_max.jpg',
+                descript: 'Zakarpattya Region',
+            },
+            {
+                src: './img/ua-21/488px-Karpatska_Ukraina-2_COA.jpg',
+                descript: 'Coat of Arms of Zakarpattya Region',
+            },
+            {
+                src: './img/ua-21/133.jpg',
+                descript: 'Zakarpattya Region',
+            },
+        ],
+    },
+
+    'UA-46': {
+        title: 'Lviv Region',
+        subtitle: 'Population is estimated as 2,512,084 (2020)',
+        text: [
+            `Lviv Region is situated in south-western outskirts of the East-European plain and western part of the Northern macroslope of the Ukrainian Carpathians. There are five natural districts on its territory such as mountain Carpathians on the south, adjoining to it Pre-Carpathians sublimity, Podil plateau in the central part, and Small Polissya and Volyn sublimity in the north.`,
+            `Lviv Region borders with five regions of Ukraine and Republic Poland. The area of Lviv Region is 21.8 thousand square km or 3.6% of general territory of Ukraine.`,
+            `Variety of natural conditions, riches of resources and convenient geographical location on crossing of the European transport corridors creates pre-conditions for subsequent development of tourism and organization of rest. And the powerful economic, research and cultural potential, providing all spheres of social-economic and spiritual life with highly skilled specialists puts Lviv Region into the leading place in Ukraine.`
+        ],
+        media: [
+            {
+                src: './img/ua-46/321.jpg',
+                descript: 'Lviv',
+            },
+            {
+                src: './img/ua-46/123.jpg',
+                descript: 'Coat of Arms of Lviv Region',
+            },
+            {
+                src: './img/ua-46/gory_1920-min.jpg',
+                descript: 'Lviv Region',
+            },
+        ],
+    },
+
+    'UA-07': {
+        title: 'Volyn region',
+        subtitle: 'Population is estimated as 1,031,421 (2020)',
+        text: [
+            `Volyn is the ancient land that is stretched out in the basin of pictiresque south inflows of the Prypyat River and in the riverheads of the Western Buh River. The land, covered with the legends of ancient times, is glorified by monuments, which remind about the fight and bravery of our ancestors. This is a region of everlasting forests, groves and valleys, blue-eyed lakes, slow and rapid rivers.`,
+            `The region has a favourable geographical position, located in the north-western part of Ukraine within the zone of Ukrainian Polissya. The area of the region makes 20.2 thousand sq. m.`,
+            `It is possible to familiarize with the rich historical and cultural inheritance of Volyn, visiting the Museum-Preserve "Cossack Graves", Memorial Complex in the village of Zaturka, created in honour of Vyacheslav Lypynsky, Museum-Farmstead of Prominent Poetess Lesya Ukrayinka in the village of Kolodyazhne, Historical-Regional Museum and Museum of Volyn Icon in Lutsk.`
+        ],
+        media: [
+            {
+                src: './img/ua-07/1325.jpg',
+                descript: 'Volyn region',
+            },
+            {
+                src: './img/ua-07/250px-Coat_of_Arms_of_Volyn_Oblast.jpg',
+                descript: 'Coat of Arms of Volyn region',
+            },
+            {
+                src: './img/ua-07/Fishing-in-Volyn-region.jpg',
+                descript: 'Volyn region',
+            },
+        ],
+    },
+}
+
+export default content
