+++ submissions/Ant-C-tech/dom-api-task/js/questions.js
@@ -1,0 +1,16 @@
+'use strict'
+
+const randomQuestions = [
+    'What do you know about this district?',
+    'Do you know something about area?',
+    'How many people live here?',
+    'Hey, common, say something!',
+    'Is this your homeland?',
+    'Have you ever been here?',
+    'Your aunt lives here?',
+    'Do you want to go there?',
+    'When are you planning to go there?',
+    'This is a wonderful place, do you want to spend here you vacation?'
+]
+
+export default randomQuestions
